server/requirements.txt
numpy==1.26.3
flask==2.2.5
flask-sqlalchemy==3.0.3
flask-cors==4.0.0
gunicorn==21.2.0
firebase-admin==6.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
finvizfinance==1.1.0
requests>=2.31.0
pandas==1.5.3
openai>=1.3.0
yfinance==0.2.33